Added separated http client from DownloadTopographyTask
